+++ R/baujat.rma.r
@@ -13,6 +13,9 @@ baujat.rma <- function(x, xlim, ylim, xlab, ylab, cex, symbol="ids", grid=TRUE, 
    if (x$k == 1)
       stop(mstyle$stop("Stopped because k = 1."))
 
+   if (exists(".darkplots") && .isTRUE(.darkplots))
+      par(fg="gray95", bg="gray10", col="gray95", col.axis="gray95", col.lab="gray95", col.main="gray95", col.sub="gray95")
+
    ### grid argument can either be a logical or a color
 
    if (is.logical(grid)) {