+++ R/funnel.default.r
@@ -23,6 +23,9 @@ label=FALSE, offset=0.4, legend=FALSE, ...) {
 
    atransf.char <- deparse(atransf)
 
+   if (exists(".darkplots") && .isTRUE(.darkplots))
+      par(fg="gray95", bg="gray10", col="gray95", col.axis="gray95", col.lab="gray95", col.main="gray95", col.sub="gray95")
+
    if (missing(back)) {
       if (.is.dark(par("bg"))) {
          back <- "gray20"