+++ R/plot.gosh.rma.r
@@ -40,6 +40,9 @@ xlim, ylim, xhist=TRUE, yhist=TRUE, hh=0.3, breaks, adjust, lwd, labels, ...) {
 
    }
 
+   if (exists(".darkplots") && .isTRUE(.darkplots))
+      par(fg="gray95", bg="gray10", col="gray95", col.axis="gray95", col.lab="gray95", col.main="gray95", col.sub="gray95")
+
    lplot  <- function(..., trim) plot(...)
    lpairs <- function(..., trim) pairs(...)