+++ R/plot.permutest.rma.uni.r
@@ -8,6 +8,9 @@ plot.permutest.rma.uni <- function(x, beta, alpha, QM=FALSE, QS=FALSE,
 
    .chkclass(class(x), must="permutest.rma.uni")
 
+   if (exists(".darkplots") && .isTRUE(.darkplots))
+      par(fg="gray95", bg="gray10", col="gray95", col.axis="gray95", col.lab="gray95", col.main="gray95", col.sub="gray95")
+
    if (missing(col)) {
       if (.is.dark(par("bg"))) {
          col <- "gray50"